+++ Disciple/Changes.md
@@ -2,6 +2,9 @@
 
 ## v1.9.2 (in development)
 Supports **Access v1.7** and **Flow v2.8**
+### New methods
+- **ResponseParseResult** (object)
+  - Added `.failedFuture(Throwable)`
 
 ## v1.9.1 - 01.11.2025
 A very small update, introducing a new response parser for file-writing.